break(iot-deps, iot-serv): Fix date time parser utility converting dates to local time when converting to strings
#1414

+++ deps/src/test/java/com/microsoft/azure/sdk/iot/deps/serializer/JobsResponseParserTest.java
@@ -44,7 +44,7 @@ public class JobsResponseParserTest
         // assert
         assertEquals("jobName", Deencapsulation.getField(jobsResponseParser, "jobId"));
         assertEquals("DeviceId IN ['new_device']", Deencapsulation.getField(jobsResponseParser, "queryCondition"));
-        assertEquals(120L, Deencapsulation.getField(jobsResponseParser, "maxExecutionTimeInSeconds"));
+        assertEquals(120L, (long) Deencapsulation.getField(jobsResponseParser, "maxExecutionTimeInSeconds"));
         assertEquals("scheduleUpdateTwin", Deencapsulation.getField(jobsResponseParser, "type").toString());
         assertEquals("enqueued", Deencapsulation.getField(jobsResponseParser, "jobsStatus").toString());
         assertEquals("Valid failure reason", Deencapsulation.getField(jobsResponseParser, "failureReason"));
